client/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "doc2pdf"
version = "0.1.0"
description = "A modern Python CLI application to convert documents to PDF"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "User", email = "user@example.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0.0",
    "pywin32>=306",
    "pypdf>=3.0.0",
    "pdfminer.six>=20221105",
]

[project.scripts]
doc2pdf = "src.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
